+++ NEWS.md
@@ -3,7 +3,7 @@ Julia v1.7 Release Notes
 
 New language features
 ---------------------
-
+* `count` and `findall` now accept an `AbstractChar` argument to search for a character in a string ([#38675]).
 
 Language changes
 ----------------